+++ DuAn1_Nhom3/src/repository/SanPhamChiTietRepository.java
@@ -28,7 +28,7 @@ public class SanPhamChiTietRepository {
             while(rs.next()) {
                 SanPhamChiTiet spct = new SanPhamChiTiet(rs.getInt(1),rs.getInt(2),rs.getInt(3),
                         rs.getInt(4),rs.getInt(5),rs.getInt(6),rs.getInt(7),
-                        rs.getInt(8),rs.getInt(9),rs.getString(10),rs.getInt(11),rs.getDouble(12),
+                        rs.getInt(8),rs.getInt(9),rs.getString(10),rs.getInt(11),rs.getBigDecimal(12),
                         rs.getString(13),rs.getInt(14),rs.getDate(15),rs.getDate(16));
                 listSPCT.add(spct);
             }
@@ -77,7 +77,7 @@ public class SanPhamChiTietRepository {
             ps = con.prepareStatement(sql);
             rs = ps.executeQuery();
             while(rs.next()) {
-                SPCTViewModel spctv = new SPCTViewModel(rs.getInt(1), rs.getInt(2), rs.getString(3), rs.getString(4), rs.getString(5), rs.getInt(6), rs.getDouble(7), rs.getInt(8));
+                SPCTViewModel spctv = new SPCTViewModel(rs.getInt(1), rs.getInt(2), rs.getString(3), rs.getString(4), rs.getString(5), rs.getInt(6), rs.getBigDecimal(7), rs.getInt(8));
                 list.add(spctv);
             }
             return list;
